refactor: move AppVersion to model package
